Clean up rename endpoint
