--- src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/EndpointContracts/IRenameEndpoint.cs
@@ -1,12 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the MIT license. See License.txt in the project root for license information.
-
-using Microsoft.CommonLanguageServerProtocol.Framework;
-using Microsoft.VisualStudio.LanguageServer.Protocol;
-
-namespace Microsoft.AspNetCore.Razor.LanguageServer.EndpointContracts;
-
-[LanguageServerEndpoint(Methods.TextDocumentRenameName)]
-internal interface IRenameEndpoint : IRazorRequestHandler<RenameParamsBridge, WorkspaceEdit?>, IRegistrationExtension
-{
-}
