--- src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/EndpointContracts/RenameParamsBridge.cs
@@ -1,10 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the MIT license. See License.txt in the project root for license information.
-
-using Microsoft.VisualStudio.LanguageServer.Protocol;
-
-namespace Microsoft.AspNetCore.Razor.LanguageServer.EndpointContracts;
-
-internal class RenameParamsBridge : RenameParams, ITextDocumentPositionParams
-{
-}
