+++ src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Refactoring/RenameEndpointTest.cs
@@ -14,7 +14,6 @@ using Microsoft.AspNetCore.Razor.Language;
 using Microsoft.AspNetCore.Razor.Language.Components;
 using Microsoft.AspNetCore.Razor.Language.Intermediate;
 using Microsoft.AspNetCore.Razor.LanguageServer.Common;
-using Microsoft.AspNetCore.Razor.LanguageServer.EndpointContracts;
 using Microsoft.AspNetCore.Razor.LanguageServer.Extensions;
 using Microsoft.AspNetCore.Razor.LanguageServer.Protocol;
 using Microsoft.AspNetCore.Razor.LanguageServer.Test;
@@ -29,7 +28,7 @@ using Moq;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace Microsoft.AspNetCore.Razor.LanguageServer.Refactoring.Test;
+namespace Microsoft.AspNetCore.Razor.LanguageServer.Refactoring;
 
 [UseExportProvider]
 public class RenameEndpointTest : LanguageServerTestBase
@@ -50,7 +49,7 @@ public class RenameEndpointTest : LanguageServerTestBase
         var languageServerFeatureOptions = Mock.Of<LanguageServerFeatureOptions>(options => options.SupportsFileManipulation == false && options.ReturnCodeActionAndRenamePathsWithPrefixedSlash == false, MockBehavior.Strict);
         var endpoint = CreateEndpoint(languageServerFeatureOptions);
         var uri = new Uri("file:///c:/First/Component1.razor");
-        var request = new RenameParamsBridge
+        var request = new RenameParams
         {
             TextDocument = new TextDocumentIdentifier
             {
@@ -74,7 +73,7 @@ public class RenameEndpointTest : LanguageServerTestBase
     {
         // Arrange
         var uri = new Uri("file:///c:/First/Component1.razor");
-        var request = new RenameParamsBridge
+        var request = new RenameParams
         {
             TextDocument = new TextDocumentIdentifier
             {
@@ -125,7 +124,7 @@ public class RenameEndpointTest : LanguageServerTestBase
     {
         // Arrange
         var uri = new Uri("file:///c:/Second/ComponentWithParam.razor");
-        var request = new RenameParamsBridge
+        var request = new RenameParams
         {
             TextDocument = new TextDocumentIdentifier
             {
@@ -149,7 +148,7 @@ public class RenameEndpointTest : LanguageServerTestBase
     {
         // Arrange
         var uri = new Uri("file:///c:/Second/ComponentWithParam.razor");
-        var request = new RenameParamsBridge
+        var request = new RenameParams
         {
             TextDocument = new TextDocumentIdentifier
             {
@@ -173,7 +172,7 @@ public class RenameEndpointTest : LanguageServerTestBase
     {
         // Arrange
         var uri = new Uri("file:///c:/Second/ComponentWithParam.razor");
-        var request = new RenameParamsBridge
+        var request = new RenameParams
         {
             TextDocument = new TextDocumentIdentifier
             {
@@ -197,7 +196,7 @@ public class RenameEndpointTest : LanguageServerTestBase
     {
         // Arrange
         var uri = new Uri("file:///c:/Second/ComponentWithParam.razor");
-        var request = new RenameParamsBridge
+        var request = new RenameParams
         {
             TextDocument = new TextDocumentIdentifier
             {
@@ -221,7 +220,7 @@ public class RenameEndpointTest : LanguageServerTestBase
     {
         // Arrange
         var uri = new Uri("file:///c:/Second/ComponentWithParam.razor");
-        var request = new RenameParamsBridge
+        var request = new RenameParams
         {
             TextDocument = new TextDocumentIdentifier
             {
@@ -245,7 +244,7 @@ public class RenameEndpointTest : LanguageServerTestBase
     {
         // Arrange
         var uri = new Uri("file:///c:/First/Index.razor");
-        var request = new RenameParamsBridge
+        var request = new RenameParams
         {
             TextDocument = new TextDocumentIdentifier
             {
@@ -318,7 +317,7 @@ public class RenameEndpointTest : LanguageServerTestBase
     {
         // Arrange
         var uri = new Uri("file:///c:/Second/Component3.razor");
-        var request = new RenameParamsBridge
+        var request = new RenameParams
         {
             TextDocument = new TextDocumentIdentifier
             {
@@ -372,7 +371,7 @@ public class RenameEndpointTest : LanguageServerTestBase
     {
         // Arrange
         var uri = new Uri("file:///c:/Dir1/Directory1.razor");
-        var request = new RenameParamsBridge
+        var request = new RenameParams
         {
             TextDocument = new TextDocumentIdentifier
             {
@@ -446,7 +445,7 @@ public class RenameEndpointTest : LanguageServerTestBase
         var endpoint = CreateEndpoint(languageServerFeatureOptions, documentMappingServiceMock.Object, languageServerMock.Object);
 
         var uri = new Uri("file:///c:/Second/ComponentWithParam.razor");
-        var request = new RenameParamsBridge
+        var request = new RenameParams
         {
             TextDocument = new TextDocumentIdentifier
             {
@@ -480,7 +479,7 @@ public class RenameEndpointTest : LanguageServerTestBase
 
         var endpoint = CreateEndpoint(languageServerFeatureOptions, documentMappingServiceMock.Object, languageServerMock.Object);
 
-        var request = new RenameParamsBridge
+        var request = new RenameParams
         {
             TextDocument = new TextDocumentIdentifier
             {
